+++ src/redux/consts/index.js
@@ -3,3 +3,4 @@
 export const CREATE_SURVEY = 'CREATE_SURVEY';
 export const UPDATE_SURVEY = 'UPDATE_SURVEY';
 export const GET_SURVEY_BY_ID = 'GET_SURVEY_BY_ID';
+export const DELETE_SURVEY = 'DELETE_SURVEY';