+++ src/views/SurveyList.jsx
@@ -1,5 +1,6 @@
 // Import libraries
 import { connect } from 'react-redux';
+import { Link } from 'react-router-dom';
 // Import components
 import SurveyCard from '../components/SurveyCard';
 
@@ -12,6 +13,16 @@ const SurveyList = ({ surveys }) => (
           <SurveyCard survey={survey} />
         </section>
       )}
+      {surveys.length === 0 &&
+        <section className="d-flex flex-column align-items-center">
+          <h5 className="text-info">There aren't any surveys, create your first survey</h5>
+          <Link to="/create">
+            <button type="button" className="btn btn-primary mt-3">
+              Create survey
+            </button>
+          </Link>
+        </section>
+      }
     </section>
   </section>
 );